+++ controllers/dns.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"context"
+	"time"
 
 	gsannotations "github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/pkg/errors"
@@ -12,6 +13,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
+	"github.com/aws-resolver-rules-operator/pkg/k8sclient"
 	"github.com/aws-resolver-rules-operator/pkg/resolver"
 )
 
@@ -97,25 +99,50 @@ func (r *DnsReconciler) reconcileNormal(ctx context.Context, awsCluster *capa.AW
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	addrType := capi.MachineExternalIP
-	if awsCluster.Annotations[gsannotations.AWSVPCMode] == gsannotations.AWSVPCModePrivate {
-		addrType = capi.MachineInternalIP
-	}
+	cluster := buildCluster(awsCluster, identity)
 
-	bastionIp, err := r.awsClusterClient.GetBastionIp(ctx, awsCluster, addrType)
-	if err != nil {
+	bastionIp, err := r.getBastionIp(ctx, awsCluster)
+	if err != nil && !errors.Is(err, &k8sclient.BastionNotFoundError{}) {
 		return ctrl.Result{}, errors.WithStack(err)
 	}
-
-	cluster := buildCluster(awsCluster, identity)
 	cluster.BastionIp = bastionIp
+
+	requeueAfter := 0 * time.Minute
+	// If there is a bastion machine, but it has no IP address just yet, we want to reconcile again soonish
+	if !errors.Is(err, &k8sclient.BastionNotFoundError{}) && bastionIp == "" {
+		requeueAfter = 1 * time.Minute
+	}
 
 	err = r.dnsZone.CreateHostedZone(ctx, logger, cluster)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
 
-	return ctrl.Result{}, nil
+	return ctrl.Result{
+		RequeueAfter: requeueAfter,
+	}, nil
+}
+
+// getBastionIp tries to find a bastion machine in this cluster and fetch its IP address from the status field.
+// It will return the internal IP address when using private VPC mode, or an external IP address otherwise.
+func (r *DnsReconciler) getBastionIp(ctx context.Context, awsCluster *capa.AWSCluster) (string, error) {
+	bastionMachine, err := r.awsClusterClient.GetBastionMachine(ctx, awsCluster.Name)
+	if err != nil {
+		return "", errors.WithStack(err)
+	}
+
+	addressType := capi.MachineExternalIP
+	if awsCluster.Annotations[gsannotations.AWSVPCMode] == gsannotations.AWSVPCModePrivate {
+		addressType = capi.MachineInternalIP
+	}
+
+	for _, addr := range bastionMachine.Status.Addresses {
+		if addr.Type == addressType {
+			return addr.Address, nil
+		}
+	}
+
+	return "", nil
 }
 
 // reconcileDelete deletes the hosted zone and the DNS records for the workload cluster.